pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "soyscope"
version = "0.1.0"
description = "Industrial Soy Uses Search & Tracking Tool"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "exa_py>=1.0.0",
    "pyalex>=0.15",
    "semanticscholar>=0.8",
    "habanero>=1.2",
    "biopython>=1.83",
    "tavily-python>=0.5",
    "httpx>=0.27",
    "aiolimiter>=1.1",
    "tenacity>=8.2",
    "diskcache>=5.6",
    "rapidfuzz>=3.5",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.0",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    "streamlit>=1.35",
    "plotly>=5.18",
    "pydantic>=2.5",
]

[project.scripts]
soyscope = "soyscope.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
